requirements.txt
tweepy
facepy
pyyaml
pyredis
pytz
selenium
flask